rpi/code/requirements.txt
adafruit-circuitpython-adxl34x
spidev
RPi.GPIO
rpi.gpio
smbus
black
Adafruit-Blinka